% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{checkModels}
\alias{checkModels}
\title{checkModels(S3 function)}
\usage{
checkModels(class_partition, class_neural, class_linear)
}
\value{
boolean
}
\description{
help function for checking, if models were created
}
