% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{create.farseer.data.frame}
\alias{create.farseer.data.frame}
\title{create (farseer.data.frame)}
\usage{
\method{create}{farseer.data.frame}(farseerDataFrame, formula, additional_targets, test)
}
\description{
Call of the create function on an farseer.data.frame object creates new farseer.models
}
