% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{create.farseer.models}
\alias{create.farseer.models}
\title{create (farseer.models)}
\usage{
create.farseer.models(
  farseerDataFrame,
  target,
  neural_max_threshold = 0.1,
  ...
)
}
\arguments{
\item{farseerDataFrame}{a data for training (must be a farseer.data.frame)}

\item{target}{an index of target from farseer.data.frame$target.variables}
}
\value{
a farseer.models object
}
\description{
creates a single set of farseer models
}
