% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{createFormula}
\alias{createFormula}
\title{Help function: create formula}
\usage{
createFormula(names)
}
\arguments{
\item{names}{colnames of a data.frame}
}
\value{
formula
}
\description{
Creates a formula using last entry in a char vector as dependant variable
and all others as independent variables.
}
