% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseer.R
\name{disclaimer}
\alias{disclaimer}
\title{disclaimer (function)
prints the disclaimer before usage of farseer or simulate}
\usage{
disclaimer()
}
\description{
disclaimer (function)
prints the disclaimer before usage of farseer or simulate
}
