% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataFrame.R
\name{factorize.data.frame}
\alias{factorize.data.frame}
\title{factorize.data.frame}
\usage{
\method{factorize}{data.frame}(obj)
}
\arguments{
\item{obj}{a dataFrame object with data to be factorized. Non-Factor data will be left unchanged.}
}
\description{
This function converts all factors to numeric values 0 and 1.
}
\details{
For every factor level save one a new variable is created {name_factor_name}
last factor level does not have to be included, as if all the values are 0 it means this case belongs to the group
}
\author{
Kornel Skitek 2020
}
