% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataSets.R
\name{factorize}
\alias{factorize}
\title{generic function for factorizing - changind factor levels into numerical 0 or 1 values}
\usage{
factorize(obj)
}
\arguments{
\item{obj}{\itemize{
\item object to be factorized
}

see factorize.data.frame for details}
}
\description{
generic function for factorizing - changind factor levels into numerical 0 or 1 values
}
