% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{farseer.bland.altmann}
\alias{farseer.bland.altmann}
\title{farseer.bland.altmann (S3 function)}
\usage{
farseer.bland.altmann(predictions, targetName = "original", title = NULL)
}
\arguments{
\item{predictions}{a data.frame containing predictions of the models and the target variable}

\item{targetName}{name of the column with measured data}

\item{title}{title of plot}
}
\description{
creates a bland-altmann plot for numeric predictions
}
