% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{farseer.models}
\alias{farseer.models}
\title{farseer.models (S3 class)}
\usage{
farseer.models(farseerDataFrame, mode = "create", ...)
}
\description{
Container class for models trained with farseer-package.
It is a list of models
\tabular{ll}{
$linear \tab linear model \cr
$partition \tab partition tree \cr
$neural \tab neural networks
}
}
