% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseer.R
\name{farseer.plot}
\alias{farseer.plot}
\title{plot.farseer (function)}
\usage{
farseer.plot(obj, plotPartition = TRUE)
}
\arguments{
\item{obj}{a farseer}

\item{plotPartition}{if TRUE, partition trees will also be plotted}
}
\description{
Plots test results of a \link{farseer} object.
}
