% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseer.R
\name{farseer.simulate}
\alias{farseer.simulate}
\title{farseer.simulate}
\usage{
farseer.simulate(
  fars,
  newData,
  simulateFor,
  modelNames = NULL,
  range = c(0, 1),
  difference = 0.1
)
}
\arguments{
\item{fars}{a farseer object.}

\item{newData}{a data.frame with data of the patients to be simulated. All variables have to be named exactly the same as in training set,
but newData can contain more variables. It is subsetted prior to evaluation.}

\item{simulateFor}{a character vector of columns, for which values should be simulated}

\item{modelNames}{character vector which models should be used for prediction. If left NULL, all available models will be used.}

\item{range}{range of values to be simulated for numeric values.}

\item{difference}{<- difference between values in range}
}
\value{
data.frame with simulated and predicted values
}
\description{
simulates all possible factor levels, and/or numerical data. In case of numerical data, values 0:1 with a difference of 0.1 will be generated,
as to simulate wide range of normalized data. This behavior can be changed by setting range and difference parameter.
}
