% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{farseer.training.vector}
\alias{farseer.training.vector}
\title{farseer.training.vector (function)}
\usage{
farseer.training.vector(noOfCases, trainingFraction = 0.9, seed = NULL)
}
\arguments{
\item{noOfCases}{integer, number of cases in a data.frame}

\item{trainingFraction}{how many cases should be a training set?}

\item{seed}{if it is not NULL, a seed for randomization is set}
}
\description{
Gets a integer vector of case numbers for the training set
}
\details{
##gets a psedo-random vector
trainingVector <- farseer.training.vector(nrow(dataFrame))

##smaller training fraction
trainingVector <- farseer.training.vector(nrow(dataFrame), trainingFraction = 0.7)

##set a seed
trainingVector <- farseer.training.vector(nrow(dataFrame), seed = 123)
}
