% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataSets.R
\name{farseerDataSet}
\alias{farseerDataSet}
\title{farseerDataSet - S3 class for creating and storing factorized and normalized datasets
constructor for a farseerDataSet}
\usage{
farseerDataSet(dataFrame)
}
\arguments{
\item{dataFrame}{a data.frame containing data for modeling, without target variables

functions: \link{factorize.data.set}}
}
\value{
list
$original: original data.frame
$factorized: all factors converted to 0 and 1
$normalized: all values are ranged 0-1. See factorize.data.frame
}
\description{
farseerDataSet - S3 class for creating and storing factorized and normalized datasets
constructor for a farseerDataSet
}
\author{
Kornel Skitek 2020
}
