% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{multiplot}
\alias{multiplot}
\title{Help function: draw multiple plots}
\usage{
multiplot(plots, formula, dataFrame)
}
\arguments{
\item{plots}{a character vector of plots to be created from c("Correlation", "Bland-Altman", "ROC")}
}
\description{
Help function: draw multiple plots
}
