% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataFrame.R
\name{normalize.data.frame}
\alias{normalize.data.frame}
\title{normalize converts all the values into values of range(0,1)}
\usage{
\method{normalize}{data.frame}(obj)
}
\arguments{
\item{obj}{data.frame factorized using \code{\link{factorize.data.frame}}}
}
\value{
data.frame
}
\description{
This allows for correct training, if the variables have very different ranges
for example:
if a antihypertension drug has a dosage range of 2 to 10 mg
and initial mean blood pressure of our patients varies from 100 to 200
the initial blood pressure will be overfitted, compared to drug dose.
}
