% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataSets.R
\name{normalize}
\alias{normalize}
\title{generic function for normalizing}
\usage{
normalize(obj)
}
\arguments{
\item{obj}{see normalize.data.frame for details}
}
\description{
generic function for normalizing
}
