% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataFrame.R
\name{normalize.vector}
\alias{normalize.vector}
\title{normalizes the given vector}
\usage{
\method{normalize}{vector}(obj)
}
\arguments{
\item{obj}{numerical vector to be normalized.}
}
\description{
normalizes the given vector
}
