% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{performance.classification}
\alias{performance.classification}
\title{performance.classification (S3 function)}
\usage{
performance.classification(
  predictions,
  trueValue,
  cuttoff = 1,
  predictionNames = c("linear", "partition", "neural")
)
}
\arguments{
\item{predictions}{a data.frame containig predictions of the models}

\item{trueValue}{a vector of the measured value}

\item{cutoff}{numeric: value for labeling, all trueValues < cuttoff will be labeled 0, otherwise 1;
non-numeric: if it is equal to cutoff, will be labeled 0, otherwise 1}
}
\value{
list \describe{
\item{predictions}{a list of \link[ROCR]{prediction-class} objects for every model}
\item{performance}{a list of \link[ROCR]{performance-class} objects for every model; performance is called with "tpr", "fpr" options}
\item{auc}{a list of \link[ROCR]{performance-class} objects for every model;performance is called with "auc" option}
}
}
\description{
calculates performance measurements of models.
}
