% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{predict.farseer.models}
\alias{predict.farseer.models}
\title{predict(farseer.models)}
\usage{
\method{predict}{farseer.models}(models, newData, test = TRUE)
}
\arguments{
\item{models}{a farseer.models object with the trained models}

\item{newData}{data to be used for predictions}

\item{test}{if true, models will be tested and plots created}
}
\value{
data.frame with predictions for every model
}
\description{
predicts data using trained models
}
