% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{rocplot}
\alias{rocplot}
\title{rocplot (S3 function)}
\usage{
rocplot(
  predictions,
  trueValue,
  cuttoff = 1,
  predictionNames = c("linear", "partition", "neural"),
  title = NULL
)
}
\arguments{
\item{predictions}{a data.frame containig predictions of the models}

\item{trueValue}{a vector of the measured value}

\item{cutoff}{numeric value for labeling, all trueValues < cuttoff will be labeled 0, otherwise 1}
}
\value{
a roc plot
}
\description{
creates a roc-plot for models
}
