% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{checkModels}
\alias{checkModels}
\title{checkModels(S3 function)}
\usage{
checkModels(class_partition, class_neural, class_linear)
}
\arguments{
\item{class_partition}{class of partition model}

\item{class_neural}{class of neural networks model}

\item{class_linear}{class of linear model}
}
\value{
boolean
}
\description{
help function for checking, if models were created. If the classes
are not as expected, FALSE is returned.
}
