% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataFrame.R
\name{farseer.data.frame}
\alias{farseer.data.frame}
\title{farseer.data.frame - S3 class for creating and storing factorized and normalized datasets}
\usage{
farseer.data.frame(dataFrame, formula, additional_targets = NULL)
}
\arguments{
\item{dataFrame}{a data.frame containing data for modeling}

\item{formula}{formula for training of \link{farseer.models}}

\item{additional_targets}{additional dependand variables, for which models should be
created using same dependand variables as in original formula.}
}
\value{
list
$data: resulting data.frame
$factorised: character vector of columns containing vectors converted to numerical values 0,1 See \code{\link{factorize.data.frame}}
$normalised: character vector of columns containing normalized numerical values See \code{\link{normalize.data.frame}}
$timestamp: date and time of creation
$denormalization: used for denormalization of data, if needed
$model.variables: character vector of model varaibles
$target.variables: character vector of target variables
}
\description{
farseer.data.frame - S3 class for creating and storing factorized and normalized datasets
}
\author{
Kornel Skitek 2020
}
