% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataSets.R
\name{farseer.data.frame}
\alias{farseer.data.frame}
\title{farseerDataSet - S3 class for creating and storing factorized and normalized datasets}
\usage{
farseer.data.frame(dataFrame, formula, additional_targets = NULL)
}
\arguments{
\item{dataFrame}{a data.frame containing data for modeling

functions: \link{factorize.data.set}}
}
\value{
list
$data: resulting data.frame
$factorised: character vector of columns containing vectors converted to numerical values 0,1 See \code{\link{factorize.data.frame}}
$normalised: character vector of columns containing normalized numerical values See \code{\link{normalize.data.frame}}
$timestamp: date and time of creation
$denormalization: used for denormalization of data, if needed
$model.variables: character vector of model varaibles
$target.variables: character vector of target variables
}
\description{
farseerDataSet - S3 class for creating and storing factorized and normalized datasets
}
\author{
Kornel Skitek 2020
}
