% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{farseer.rocplot}
\alias{farseer.rocplot}
\title{farseer.rocplot (S3 function)}
\usage{
farseer.rocplot(performanceList, targetName, title = NULL)
}
\arguments{
\item{predictions}{a data.frame containig predictions of the models}

\item{cutoff}{numeric value for labeling, all trueValues < cuttoff will be labeled 0, otherwise 1}
}
\value{
a roc plot
}
\description{
creates a roc-plot for models
}
