% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerUtilities.R
\name{farseer.training.vector}
\alias{farseer.training.vector}
\title{farseer.training.vector (function)}
\usage{
##gets a psedo-random vector
trainingVector <- farseer.training.vector(nrows(dataFrame))

##smaller training fraction
trainingVector <- farseer.training.vector(nrows(dataFrame), trainingFraction = 0.7)

##set a seed
trainingVector <- farseer.training.vector(nrows(dataFrame), seed = 123)
}
\arguments{
\item{noOfCases}{integer, number of cases in a data.frame}

\item{trainingFraction}{how many cases should be a training set?}

\item{seed}{if it is not NULL, a seed for randomization is set}
}
\description{
Gets a integer vector of case numbers for the training set
}
\examples{
#create a farseer.data.frame
data <- farseer.data.frame(formula = formula, dataFrame)
#get the vector
training.vector <- farseer.training.vector(noOfCases = nrows(data$data))

#get the sets
training.set <- data$data[training.vector,]
testing.set <- data$data[-training.vector,]

}
