% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataSets.R
\name{models.farseer.data.frame}
\alias{models.farseer.data.frame}
\title{create (farseer.data.frame)}
\usage{
\method{models}{farseer.data.frame}(farseerDataFrame, ...)
}
\arguments{
\item{farseerDataFrame}{prepared data frame for training}
}
\value{
list of models
}
\description{
Call of the create function on an farseer.data.frame object creates new farseer.models
}
