% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerDataSets.R
\name{normalize.vector}
\alias{normalize.vector}
\title{normalizes the given vector}
\usage{
\method{normalize}{vector}(x)
}
\arguments{
\item{x}{numerical vector to be normalized.}
}
\description{
normalizes the given vector
}
