% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farseerModels.R
\name{performance.numeric}
\alias{performance.numeric}
\title{performance.numeric (S3 function)}
\usage{
performance.numeric(
  predictions,
  trueValue,
  predictionNames = c("linear", "partition", "neural")
)
}
\arguments{
\item{predictions}{predicted values}

\item{trueValue}{observed values}

\item{predictionNames}{names of the columns, where predictions can be found}
}
\value{
list: \describe{
\item{errors}{data.frame with measured, predicted, absolute and relative errors for all models}
\item{performance}{data.frame with correlation, mean error and SD of mean error}
}
}
\description{
Calculates the absolute and relative errors of predictions, as well as their means and SDs.
Additionally calculates the correlations.
}
